+++ app/database/migrations/2014_08_10_120843_create_projects_table.php
@@ -11,16 +11,21 @@ class CreateProjectsTable extends Migration {
 	 * @return void
 	 */
 	public function up()
-	{
-            Schema::create('projects', function(Blueprint $table)
+        {
+            if (!Schema::hasTable('projects'))
             {
-                //Create initial projects table 
-                $table->increments('id');
-                $table->integer('user_id');
-                $table->integer('sequence');
-                $table->integer('parent_project')->nullable();
-                $table->timestamps();
-            });
+                Schema::create('projects', function(Blueprint $table)
+                {
+                    //Create initial projects table 
+                    $table->increments('id');
+                    $table->integer('user_id')->unsigned();
+                    $table->foreign('user_id')->references('id')->on('users');
+                    $table->integer('parent_project_id')->unsigned()->nullable();
+                    $table->foreign('parent_project_id')->references('id')->on('projects');
+                    $table->integer('sequence');
+                    $table->timestamps();
+                });
+            }
 	}
 
 	/**
@@ -30,8 +35,11 @@ class CreateProjectsTable extends Migration {
 	 */
 	public function down()
 	{
-            //Drop projects table
-            Schema::drop('projects');
+            if (Schema::hasTable('projects'))
+            {
+                //Drop projects table
+                Schema::drop('projects');
+            }
 	}
 
 }
